+++ src/calibration.rs
@@ -17,9 +17,9 @@ pub struct Calibration {
 
     pub h1 : u8,
     pub h2 : u16,
-    pub h3 : u16,
+    pub h3 : u8,
     pub h4 : u16,
     pub h5 : u16,
-    pub h6 : u16,
+    pub h6 : i8,
     pub h7 : u16    
 }
